+++ src/cisco_gnmi/xe.py
@@ -31,6 +31,7 @@ from six import string_types
 from .client import Client, proto, util
 
 logger = logging.getLogger(__name__)
+logger.setLevel(logging.DEBUG)
 
 
 class XEClient(Client):
@@ -78,99 +79,6 @@ class XEClient(Client):
     ...
     >>> delete_response = client.delete_xpaths('/Cisco-IOS-XE-native:native/hostname')
     """
-
-    def xpath_to_path_elem(self, request):
-        """Helper function for Cisco XE clients.
-
-        Parameters
-        ---------
-        request: dict containing request namespace and nodes to be worked on.
-            namespace: dict of <prefix>: <namespace>
-            nodes: list of dict
-                  <xpath>: Xpath pointing to resource
-                  <value>: value to set resource to
-                  <edit-op>: equivelant NETCONF edit-config operation
-
-        Returns
-        -------
-        tuple: namespace_modules, message dict, origin
-            namespace_modules: dict of <prefix>: <module name>
-                Needed for future support.
-            message dict: 4 lists containing possible updates, replaces,
-                deletes, or gets derived form input nodes.
-            origin str: openconfig if detected (XE uses rfc7951 by default)
-        """
-        paths = []
-        message = {
-            'update': [],
-            'replace': [],
-            'delete': [],
-            'get': [],
-        }
-        if 'nodes' not in request:
-            # TODO: raw rpc?
-            return paths
-        else:
-            namespace_modules = {}
-            origin = None
-            for prefix, nspace in request.get('namespace', {}).items():
-                module = ''
-                if '/Cisco-IOS-' in nspace:
-                    module = nspace[nspace.rfind('/') + 1:]
-                elif '/openconfig.net' in nspace:
-                    module = 'openconfig-'
-                    module += nspace[nspace.rfind('/') + 1:]
-                elif 'urn:ietf:params:xml:ns:yang:' in nspace:
-                    module = nspace.replace(
-                        'urn:ietf:params:xml:ns:yang:', '')
-                if module:
-                    namespace_modules[prefix] = module
-            for node in request.get('nodes', []):
-                if 'xpath' not in node:
-                    logger.error('Xpath is not in message')
-                else:
-                    xpath = node['xpath']
-                    value = node.get('value', '')
-                    edit_op = node.get('edit-op', '')
-
-                    for pfx, ns in namespace_modules.items():
-                        if pfx in xpath and 'openconfig' in ns:
-                            origin = 'openconfig'
-                        xpath = xpath.replace(pfx + ':', '')
-                        if isinstance(value, string_types):
-                            value = value.replace(pfx + ':', '')
-                    if edit_op:
-                        if edit_op in ['create', 'merge', 'replace']:
-                            xpath_lst = xpath.split('/')
-                            name = xpath_lst.pop()
-                            xpath = '/'.join(xpath_lst)
-                            if edit_op == 'replace':
-                                if not message['replace']:
-                                    message['replace'] = [{
-                                        xpath: {name: value}
-                                    }]
-                                else:
-                                    message['replace'].append(
-                                        {xpath: {name: value}}
-                                    )
-                            else:
-                                if not message['update']:
-                                    message['update'] = [{
-                                        xpath: {name: value}
-                                    }]
-                                else:
-                                    message['update'].append(
-                                        {xpath: {name: value}}
-                                    )
-                        elif edit_op in ['delete', 'remove']:
-                            if message['delete']:
-                                message['delete'].add(xpath)
-                            else:
-                                message['delete'] = set(xpath)
-                    else:
-                        message['get'].append(xpath)
-        return namespace_modules, message, origin
-
     def delete_xpaths(self, xpaths, prefix=None):
         """A convenience wrapper for set() which constructs Paths from supplied xpaths
         to be passed to set() as the delete parameter.
@@ -315,7 +223,7 @@ class XEClient(Client):
             json_ietf=json_ietf
         )
         for update in updates + replaces:
-            logger.info('\nGNMI set:\n{0}\n{1}'.format(9 * '=', str(update)))
+            logger.debug('\nGNMI set:\n{0}\n{1}'.format(9 * '=', str(update)))
 
         return self.set(updates=updates, replaces=replaces)
 
@@ -357,7 +265,7 @@ class XEClient(Client):
             raise Exception(
                 "xpaths must be a single xpath string or iterable of xpath strings!"
             )
-        logger.info('GNMI get:\n{0}\n{1}'.format(9 * '=', str(gnmi_path)))
+        logger.debug('GNMI get:\n{0}\n{1}'.format(9 * '=', str(gnmi_path)))
         return self.get(gnmi_path, data_type=data_type, encoding=encoding)
 
     def subscribe_xpaths(
@@ -365,7 +273,7 @@ class XEClient(Client):
         xpath_subscriptions,
         request_mode="STREAM",
         sub_mode="SAMPLE",
-        encoding="PROTO",
+        encoding="JSON_IETF",
         sample_interval=Client._NS_IN_S * 10,
         origin='openconfig'
     ):
@@ -396,7 +304,7 @@ class XEClient(Client):
             [ON_CHANGE, SAMPLE]
         encoding : proto.gnmi_pb2.Encoding, optional
             A member of the proto.gnmi_pb2.Encoding enum specifying desired encoding of returned data
-            [JSON, PROTO]
+            [JSON, JSON_IETF]
         sample_interval : int, optional
             Default nanoseconds for sample to occur.
             Defaults to 10 seconds.
@@ -470,7 +378,7 @@ class XEClient(Client):
                 raise Exception("xpath in list must be xpath or dict/Path!")
             subscriptions.append(subscription)
         subscription_list.subscription.extend(subscriptions)
-        logger.info('GNMI subscribe:\n{0}\n{1}'.format(
+        logger.debug('GNMI subscribe:\n{0}\n{1}'.format(
             15 * '=', str(subscription_list))
         )
         return self.subscribe([subscription_list])
@@ -487,3 +395,34 @@ class XEClient(Client):
             else:
                 origin = "rfc7951"
         return super(XEClient, self).parse_xpath_to_gnmi_path(xpath, origin)
+
+    def xpath_to_path_elem(self, request):
+        """Convert XML Path Language 1.0 formed xpath to gNMI PathElement.
+
+        Modeled after NETCONF Xpaths RFC 6020.
+
+        References:
+        * https://www.w3.org/TR/1999/REC-xpath-19991116/#location-paths
+        * https://www.w3.org/TR/1999/REC-xpath-19991116/#path-abbrev
+        * https://tools.ietf.org/html/rfc6020#section-6.4
+        * https://tools.ietf.org/html/rfc6020#section-9.13
+
+        Parameters
+        ---------
+        request: dict containing request namespace and nodes to be worked on.
+            namespace: dict of <prefix>: <namespace>
+            nodes: list of dict
+                  <xpath>: Xpath pointing to resource
+                  <value>: value to set resource to
+                  <edit-op>: equivelant NETCONF edit-config operation
+
+        Returns
+        -------
+        tuple: namespace_modules, message dict, origin
+            namespace_modules: dict of <prefix>: <module name>
+                Needed for future support.
+            message dict: 4 lists containing possible updates, replaces,
+                deletes, or gets derived form input nodes.
+            origin str: DME, device, or openconfig
+        """
+        return super(XEClient, self).xml_path_to_path_elem(request)
